Fixes for 64-bit build under Mingw-w64.

+++ Engine/Core/ProcessUtils.cpp
@@ -144,13 +144,13 @@ void OpenConsoleWindow()
     AllocConsole();
 
     HANDLE hOut = GetStdHandle(STD_OUTPUT_HANDLE);
-    int hCrt = _open_osfhandle((long)hOut, _O_TEXT);
+    int hCrt = _open_osfhandle((size_t)hOut, _O_TEXT);
     FILE* outFile = _fdopen(hCrt, "w");
     setvbuf(outFile, NULL, _IONBF, 1);
     *stdout = *outFile;
 
     HANDLE hIn = GetStdHandle(STD_INPUT_HANDLE);
-    hCrt = _open_osfhandle((long)hIn, _O_TEXT);
+    hCrt = _open_osfhandle((size_t)hIn, _O_TEXT);
     FILE* inFile = _fdopen(hCrt, "r");
     setvbuf(inFile, NULL, _IONBF, 128);
     *stdin = *inFile;
